backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
asyncpg==0.29.0
geoalchemy2==0.14.1
pydantic==2.5.0
python-dotenv==1.0.0
httpx==0.25.2
redis==5.0.1
geopy==2.4.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pytest==7.4.3
pytest-asyncio==0.21.1
